# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Authentication, Company, Job, Statistics, and User API
  description: API for managing user authentication, company records, job postings, statistical data, and user interactions.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/jobs/:
    get:
      tags:
        - Jobs
      summary: List Jobs
      description: Retrieve a list of jobs with pagination and sorting.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number.
        - in: query
          name: size
          schema:
            type: integer
            default: 20
          description: Number of items per page.
        - in: query
          name: sort_by
          schema:
            type: string
            default: "id"
          description: Field to sort by.
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
            default: "desc"
          description: Sort order.
      responses:
        200:
          description: List of jobs retrieved successfully.
        500:
          description: Internal server error.
    post:
      tags:
        - Jobs
      summary: "Add a Job"
      description: "Create a new job posting."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: "Title of the job."
                link:
                  type: string
                  description: "Link to the job posting."
                career_condition:
                  type: string
                  description: "Career condition required for the job."
                education:
                  type: string
                  description: "Education level required for the job."
                deadline:
                  type: string
                  description: "Application deadline for the job."
                job_sector:
                  type: string
                  description: "Sector of the job."
                company:
                  type: integer
                  description: "Company ID (required for admin users)."
                tech_ids:
                  type: array
                  items:
                    type: integer
                  description: "List of technology IDs associated with the job."
                location_ids:
                  type: array
                  items:
                    type: integer
                  description: "List of location IDs associated with the job."
      responses:
        201:
          description: "Job created successfully."
        400:
          description: "Validation error."
        500:
          description: "Internal server error."
  /api/jobs/{job_id}:
    get:
      tags:
        - Jobs
      summary: Get Job Details
      description: Retrieves detailed information for a specific job.
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: The ID of the job to retrieve details for.
      responses:
        200:
          description: Job details returned successfully.
        404:
          description: Job not found.
        500:
          description: Internal server error.
    put:
      tags:
        - Jobs
      summary: "Update Job"
      description: "Updates an existing job posting."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: "The ID of the job to update."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                company:
                  type: integer
                creator:
                  type: integer
                career_condition:
                  type: string
                education:
                  type: string
                deadline:
                  type: string
                job_sector:
                  type: string
                tech_ids:
                  type: array
                  items:
                    type: integer
                  description: "List of technology IDs to associate with the job."
                location_ids:
                  type: array
                  items:
                    type: integer
                  description: "List of location IDs to associate with the job."
      responses:
        200:
          description: "Job updated successfully."
        400:
          description: "Validation error."
        500:
          description: "Internal server error."
    delete:
      tags:
        - Jobs
      summary: "Delete Job"
      description: "Deletes a job posting."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: "The ID of the job to delete."
      responses:
        200:
          description: "Job deleted successfully."
        404:
          description: "Job not found."
        500:
          description: "Internal server error."
  /api/jobs/search:
    get:
      tags:
        - Jobs
      summary: Search Jobs
      description: Searches and filters job postings based on various criteria.
      parameters:
        - in: query
          name: keyword
          schema:
            type: string
          description: Search keyword for job titles or descriptions.
        - in: query
          name: location
          schema:
            type: string
          description: Location filter.
        - in: query
          name: tech
          schema:
            type: string
          description: Technology filter.
        - in: query
          name: career_condition
          schema:
            type: string
          description: Career condition filter.
      responses:
        200:
          description: Search results returned successfully.
        500:
          description: Internal server error.
  /api/jobs/{job_id}/applications:
    get:
      tags:
        - Job Applications
      summary: "List Applications for a Job"
      description: >
        Retrieves a list of applications for a specific job. 
        Only accessible by the job creator.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
          description: "The ID of the job to retrieve applications for."
      responses:
        200:
          description: "List of applications retrieved successfully."
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    user:
                      type: integer
                    email:
                      type: string
                    role:
                      type: string
                      enum: ["admin", "employer", "applicant"]
                    content:
                      type: string
                      description: "The content of the application."
                    created_at:
                      type: string
                      format: date-time
        403:
          description: "Permission denied."
        404:
          description: "Job not found or no applications."
        500:
          description: "Internal server error."
...
  /api/stats/jobs:
    get:
      tags:
        - Statistics
      summary: Job Application Statistics
      description: Retrieves the number of applications for each job posting.
      responses:
        200:
          description: Job application statistics retrieved successfully.
        500:
          description: Internal server error.
...